+++ sonar-plugin-server/src/main/java/jetbrains/buildserver/sonarplugin/sqrunner/manager/projectsettings/XMLBasedSQSInfo.java
@@ -1,7 +1,6 @@
-package jetbrains.buildserver.sonarplugin.sqrunner.manager;
+package jetbrains.buildserver.sonarplugin.sqrunner.manager.projectsettings;
 
 import jetbrains.buildServer.XmlStorable;
-import jetbrains.buildServer.controllers.BasePropertiesBean;
 import jetbrains.buildServer.serverSide.crypt.EncryptUtil;
 import jetbrains.buildserver.sonarplugin.Util;
 import org.jdom.Element;
@@ -13,18 +12,7 @@ import org.jetbrains.annotations.Nullable;
  *
  * SQSInfo based on XML definition
  */
-public class XMLBasedSQSInfo extends BasePropertiesBean implements SQSInfo, XmlStorable {
-    public static final String ID = "id";
-    public static final String JDBC_URL = "jdbcUrl";
-    public static final String JDBC_USERNAME = "jdbcUsername";
-    public static final String JDBC_PASSWORD = "jdbcPassword";
-    public static final String URL = "url";
-    public static final String LOGIN = "login";
-    public static final String PASSWORD = "password";
-    public static final String NAME = "name";
-    public static final String DESCRIPTION = "description";
-    private static final String[] OPEN_FIELDS = new String[] {ID, JDBC_URL, JDBC_USERNAME, URL, LOGIN, NAME, DESCRIPTION};
-    private static final String[] ENCRYPTED_FIELDS = new String[] {JDBC_PASSWORD, PASSWORD};
+public class XMLBasedSQSInfo extends BaseSQSInfo implements XmlStorable {
 
     public XMLBasedSQSInfo() {
         super(null);
@@ -38,64 +26,7 @@ public class XMLBasedSQSInfo extends BasePropertiesBean implements SQSInfo, XmlS
                            @Nullable final String jdbcUrl,
                            @Nullable final String jdbcUsername,
                            @Nullable final String jdbcPassword) {
-        super(null);
-        setProperty(ID, id);
-        setProperty(NAME, name);
-        setProperty(URL, url);
-        setProperty(LOGIN, login);
-        setProperty(PASSWORD, password);
-        setProperty(JDBC_PASSWORD, jdbcPassword);
-        setProperty(JDBC_URL, jdbcUrl);
-        setProperty(JDBC_USERNAME, jdbcUsername);
-    }
-
-    private String get(final String key) {
-        return getProperties().get(key);
-    }
-
-    @Nullable
-    public String getUrl() {
-        return get(URL);
-    }
-
-    @Nullable
-    public String getLogin() {
-        return get(LOGIN);
-    }
-
-    @Nullable
-    public String getPassword() {
-        return get(PASSWORD);
-    }
-
-    @Nullable
-    public String getJDBCUrl() {
-        return get(JDBC_URL);
-    }
-
-    @Nullable
-    public String getJDBCUsername() {
-        return get(JDBC_USERNAME);
-    }
-
-    @Nullable
-    public String getJDBCPassword() {
-        return get(JDBC_PASSWORD);
-    }
-
-    @NotNull
-    public String getId() {
-        return get(ID);
-    }
-
-    @Nullable
-    public String getName() {
-        return get(NAME);
-    }
-
-    @Nullable
-    public String getDescription() {
-        return get(DESCRIPTION);
+        super(id, name, url, login, password, jdbcUrl, jdbcUsername, jdbcPassword);
     }
 
     public void readFrom(final Element element) {
