+++ sonar-plugin-server/src/main/resources/META-INF/build-server-plugin-sonar-plugin.xml
@@ -5,7 +5,7 @@
     <bean class="jetbrains.buildserver.sonarplugin.sqrunner.BuildSummaryLinkExtension"/>
     <bean class="jetbrains.buildserver.sonarplugin.sqrunner.ServerManagementProjectTab"/>
     <bean class="jetbrains.buildserver.sonarplugin.sqrunner.ManageSQSActionController"/>
-    <bean class="jetbrains.buildserver.sonarplugin.sqrunner.manager.SQSManagerImpl"/>
+    <bean class="jetbrains.buildserver.sonarplugin.sqrunner.manager.projectsettings.SQSManagerImpl"/>
     <bean class="jetbrains.buildserver.sonarplugin.sqrunner.EditSQRRunType"/>
     <bean class="jetbrains.buildserver.sonarplugin.sqrunner.SQSPropertiesProvider"/>
     <bean class="jetbrains.buildserver.sonarplugin.sqrunner.SQRPasswordProvider"/>
